Add a PartitionTableNode to graph
Currently we don't have any synchronisation point for the partition
creation.  This means we have the "have I been called" workarounds in
the Partitioning object as a result of each PartitionNode calling into
it (see I35c5a473509f17a70270a2cbf5bf579faaeb123a for where this went
wrong).

This inserts a single PartitionTableNode at the root of all
PartitionNode objects in the graph.  We create the partition table and
attach/detach the partitions within this node.  PartitionNode thus
just becomes a place-holder for dependencies (i.e. the
create/umount/cleanup commands are blank).

Change-Id: I1b2af14791c5f6f65e52d05fb2724970cf75fb15